+++ Tests/InstanaAgentTests/Configuration/InstanaSessionTests.swift
@@ -30,6 +30,8 @@ class InstanaSessionTests: InstanaTestCase {
         let configUsi = InstanaConfiguration(reportingURL: .random, key: "KEY", httpCaptureConfig: .automatic,
                                              enableCrashReporting: false, slowSendInterval: 0.0,
                                              usiRefreshTimeIntervalInHrs: usiTrackingNotAllowed,
+                                             rateLimits: RateLimits.DEFAULT_LIMITS,
+                                             perfConfig: nil,
                                              hybridAgentId: nil,
                                              hybridAgentVersion: nil)
         let sut = InstanaSession(configuration: configUsi, propertyHandler: propertyHandler, collectionEnabled: true)
@@ -47,6 +49,8 @@ class InstanaSessionTests: InstanaTestCase {
         let configUsi = InstanaConfiguration(reportingURL: .random, key: "KEY", httpCaptureConfig: .automatic,
                                              enableCrashReporting: false, slowSendInterval: 0.0,
                                              usiRefreshTimeIntervalInHrs: (1.0 / 3600.0),
+                                             rateLimits: RateLimits.DEFAULT_LIMITS,
+                                             perfConfig: nil,
                                              hybridAgentId: nil,
                                              hybridAgentVersion: nil)
         let sut = InstanaSession(configuration: configUsi, propertyHandler: propertyHandler, collectionEnabled: true)