+++ Sources/InstanaAgent/Monitors/Performance/LowMemoryMonitor.swift
@@ -6,7 +6,7 @@ import Foundation
 #if os(tvOS) || os(watchOS) || os(iOS)
     import UIKit
 #endif
-import MachO
+import os
 
 class LowMemoryMonitor {
     let reporter: Reporter
@@ -22,43 +22,13 @@ class LowMemoryMonitor {
 
     @objc func onLowMemoryWarning(notification: Notification) {
         let unit = UInt64(1024 * 1024)
-        let usedMemory = getUsedMemory()
-        let freeMemory = getFreeMemory()
-        let maxMemory: UInt64 = ProcessInfo.processInfo.physicalMemory
+        var freeMemory: UInt64?
+        if #available(iOS 13.0, *) {
+            freeMemory = UInt64(os_proc_available_memory()) / unit
+        }
+        let maxMemory: UInt64 = ProcessInfo.processInfo.physicalMemory / unit
+
         // in mega bytes
-        reporter.submit(PerfLowMemoryBeacon(usedMemory: usedMemory != nil ? usedMemory! / unit : nil,
-                                            availableMemory: freeMemory != nil ? freeMemory! / unit : nil,
-                                            maximumMemory: maxMemory / unit))
-    }
-
-    func getUsedMemory() -> UInt64? {
-        var taskVMInfo = task_vm_info_data_t()
-        var vmSize = mach_msg_type_number_t(MemoryLayout<task_vm_info>.size) / 4
-        let kernResult: kern_return_t = withUnsafeMutablePointer(to: &taskVMInfo) {
-            $0.withMemoryRebound(to: integer_t.self, capacity: 1) {
-                task_info(mach_task_self_, task_flavor_t(TASK_VM_INFO), $0, &vmSize)
-            }
-        }
-        if kernResult == KERN_SUCCESS {
-            return UInt64(taskVMInfo.phys_footprint)
-        }
-        return nil
-    }
-
-    func getFreeMemory() -> UInt64? {
-        var vmStats = vm_statistics64()
-        var vmSize = mach_msg_type_number_t(MemoryLayout<vm_statistics64>.size) / 4
-
-        let machHost = mach_host_self()
-        let ret = withUnsafeMutablePointer(to: &vmStats) {
-            $0.withMemoryRebound(to: integer_t.self, capacity: Int(vmSize)) {
-                host_statistics64(machHost, HOST_VM_INFO64, $0, &vmSize)
-            }
-        }
-
-        if ret == KERN_SUCCESS {
-            return UInt64(vmStats.free_count) * UInt64(vm_kernel_page_size)
-        }
-        return nil
+        reporter.submit(PerfLowMemoryBeacon(availableMemory: freeMemory, maximumMemory: maxMemory))
     }
 }